Refactor: rename "token change" to "price alert"
