+++ packages/core-mobile/app/services/notifications/priceAlert/unsubscribeForPriceAlert.ts
@@ -1,9 +1,9 @@
-import { setTokenChangeSubscriptions } from 'services/notifications/tokenChange/setTokenChangeSubscriptions'
+import { setPriceAlertSubscriptions } from 'services/notifications/priceAlert/setPriceAlertSubscriptions'
 import FCMService from 'services/fcm/FCMService'
 import { registerDeviceToNotificationSender } from 'services/notifications/registerDeviceToNotificationSender'
 
-export async function unsubscribeForTokenChange(): Promise<void> {
+export async function unsubscribeForPriceAlert(): Promise<void> {
   const fcmToken = await FCMService.getFCMToken()
   const deviceArn = await registerDeviceToNotificationSender(fcmToken)
-  await setTokenChangeSubscriptions({ tokens: [], deviceArn })
+  await setPriceAlertSubscriptions({ tokens: [], deviceArn })
 }
