+++ packages/core-mobile/app/store/notifications/listeners/unsubscribeAllNotifications.ts
@@ -4,7 +4,7 @@ import { unSubscribeForBalanceChange } from 'services/notifications/balanceChang
 import { unSubscribeForNews } from 'services/notifications/news/unsubscribeForNews'
 import messaging from '@react-native-firebase/messaging'
 import Logger from 'utils/Logger'
-import { unsubscribeForTokenChange } from 'services/notifications/tokenChange/unsubscribeForTokenChange'
+import { unsubscribeForPriceAlert } from 'services/notifications/priceAlert/unsubscribeForPriceAlert'
 
 export async function unsubscribeAllNotifications(): Promise<void> {
   const fcmToken = await FCMService.getFCMToken()
@@ -15,7 +15,7 @@ export async function unsubscribeAllNotifications(): Promise<void> {
       deviceArn,
       channelIds: []
     }),
-    unsubscribeForTokenChange()
+    unsubscribeForPriceAlert()
   ]).catch(error => {
     //as fallback invalidate token so user doesn't get notifications
     messaging().deleteToken()
